+++ lib/runners/processor/eslint.rb
@@ -38,6 +38,8 @@ module Runners
     def self.config_example
       <<~'YAML'
         root_dir: project/
+        dependencies:
+          - my-eslint-plugin@2
         npm_install: false
         target:
           - src/