+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Misc:
 - Secure `pip install` [#2194](https://github.com/sider/runners/pull/2194)
 - Introduce common option `target` [#2191](https://github.com/sider/runners/pull/2191)
 - **PHPMD** Handle syntax error as issue [#2201](https://github.com/sider/runners/pull/2201)
+- Add `linter.<id>.dependencies` option for npm tools [#2202](https://github.com/sider/runners/pull/2202)
 
 ## 0.45.0